requirements.txt
biopython==1.78
Flask==2.3.2
Flask_Cors==4.0.0
gunicorn==21.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
docopt==0.6.2
idna==3.4
packaging==23.1
pipreqs==0.4.13
requests==2.31.0
urllib3==2.0.4
yarg==0.1.9
